updated matchup-location (#49)
* merged with latest updates
* main edits by Dale: Added a section on how to get the data from the ATN
Binned the multiple daily rows into a single row for each day. There is no reason to do multiple downloads for each day since our satellite data is daily or monthly. This has the added benefit of reducing the number of data requests from over 800 to under 200.
Put a loop counter in the manual download section so that students can follow the download progress.
Made a few edits to the text to make the exercise more ATN centric
* renamed file names
* added images